refactor: fiscal year を utils にまとめる
