+++ workspaces/client/app/hooks/use-auth.ts
@@ -1,16 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
+import { getFiscalYearStartDate } from "~/utils/fiscal-year";
 import { useRepository } from "./use-repository";
-
-// 今年度の最初の日付を取得する
-const getFiscalYearStartDate = () => {
-	const date = new Date();
-	const year = date.getFullYear();
-	const month = date.getMonth() + 1;
-	if (month >= 4) {
-		return new Date(year, 3, 1);
-	}
-	return new Date(year - 1, 3, 1);
-};
 
 export function useAuth() {
 	const { authRepository } = useRepository();
